requirements.txt
Django==3.2.2
djangorestframework==3.12.4
drf-yasg==1.20.0
django-cors-headers==3.7.0
pytest==8.3.2
pytest-django==4.8.0
pytest-cov==5.0.0
djoser